Reduce redundant code when testing tools
