+++ tests/phpunit/unit-tests/admin/tools/class-llms-test-admin-tool-install-forms.php
@@ -6,42 +6,19 @@
  *
  * @group admin
  * @group admin_tools
+ * @group install_forms
  *
  * @since 5.0.0
+ * @since [version] Use `LLMS_Admin_Tool_Test_Case` and remove redundant methods/tests.
  */
-class LLMS_Test_Admin_Tool_Install_Forms extends LLMS_UnitTestCase {
+class LLMS_Test_Admin_Tool_Install_Forms extends LLMS_Admin_Tool_Test_Case {
 
 	/**
-	 * Setup before class
+	 * Name of the class being tested.
 	 *
-	 * Include abstract class.
-	 *
-	 * @since 5.0.0
-	 *
-	 * @return void
+	 * @var sting
 	 */
-	public static function setUpBeforeClass() {
-
-		parent::setUpBeforeClass();
-
-		require_once LLMS_PLUGIN_DIR . 'includes/abstracts/llms-abstract-admin-tool.php';
-		require_once LLMS_PLUGIN_DIR . 'includes/admin/tools/class-llms-admin-tool-install-forms.php';
-
-	}
-
-	/**
-	 * Setup the test case
-	 *
-	 * @since 5.0.0
-	 *
-	 * @return void
-	 */
-	public function setUp() {
-
-		parent::setUp();
-		$this->main = new LLMS_Admin_Tool_Install_Forms();
-
-	}
+	const CLASS_NAME = 'LLMS_Admin_Tool_Install_Forms';
 
 	/**
 	 * Retrieve a list of core reusable block post ids
@@ -72,51 +49,6 @@ class LLMS_Test_Admin_Tool_Install_Forms extends LLMS_UnitTestCase {
 
 		$forms = new WP_Query( array( 'post_type' => 'llms_form' ) );
 		return $forms->posts;
-
-	}
-
-	/**
-	 * Test get_description()
-	 *
-	 * @since 5.0.0
-	 *
-	 * @return void
-	 */
-	public function test_get_description() {
-
-		$res = LLMS_Unit_Test_Util::call_method( $this->main, 'get_description' );
-		$this->assertTrue( ! empty( $res ) );
-		$this->assertTrue( is_string( $res ) );
-
-	}
-
-	/**
-	 * Test get_label()
-	 *
-	 * @since 5.0.0
-	 *
-	 * @return void
-	 */
-	public function test_get_label() {
-
-		$res = LLMS_Unit_Test_Util::call_method( $this->main, 'get_label' );
-		$this->assertTrue( ! empty( $res ) );
-		$this->assertTrue( is_string( $res ) );
-
-	}
-
-	/**
-	 * Test get_text()
-	 *
-	 * @since 5.0.0
-	 *
-	 * @return void
-	 */
-	public function test_get_text() {
-
-		$res = LLMS_Unit_Test_Util::call_method( $this->main, 'get_text' );
-		$this->assertTrue( ! empty( $res ) );
-		$this->assertTrue( is_string( $res ) );
 
 	}
